requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
google-api-python-client==2.169.0
google-auth-oauthlib==1.2.2
google-cloud-firestore==2.21.0
firebase-admin==7.1.0
python-dateutil==2.9.0.post0
openai>=1.0.0
